Added two more macros namely LOG_ENTRY_DBG and LOG_EXIT_DBG
Sometimes it is desired that we need function/method/code block
entry-exit logs but only if, we are debugging. In release and/or
production code we really don't want to pollute the logs with
entry-exit messages (specially for innocent ones). Thus added
two new macros which would only print the entry-exit logs if
the app is running in DEBUG mode only.